Create a shared widgets directory and move the Button component there
